Move repositories te belonged package.
